move user-groups and network-users to enterprise controllers
